# Clarify compat vs enhanced modes