## Summary
- expose a `datason.json` package that directly re-exports the stdlib module
- add `save`/`save_string` helpers and document breaking differences for enhanced loads
- document Compat vs Enhanced usage in new Two Modes guide and link from navigation

## Testing
- `pre-commit run --files README.md datason/__init__.py datason/api.py datason/json/__init__.py docs/api/modern-deserialization.md docs/api/modern-serialization.md docs/two-modes.md mkdocs.yml tests/unit/test_json_parity.py scripts/simple_json_check.py`
- `pytest tests/unit/test_json_parity.py tests/unit/test_api_json_replacement_coverage.py`


------
https://chatgpt.com/codex/tasks/task_e_6895376b9eb48325bf6586df5cb09879