@danielendler 2025-08-07
# 🚀 DataSON PR Performance Analysis

**PR #74** | **Commit**: `ab951967cd6ead6e4142b7a13a4fb7f2a498aa42`

## 📊 Benchmark Results

**Suite**: pr_optimized | **Tests Run**: 5 | **Success Rate**: 100.0%

### 🎯 DataSON Performance Summary

| Metric | Result | Benchmark Details | Status |
|--------|--------|-------------------|--------|
| Serialization (avg) | **1.131 ms** | 5 measurements across 5 scenarios | ✅ |
| Deserialization (avg) | **0.000 ms** | 0 measurements | ❌ |
| Success Rate | **100.0%** | 5 scenarios tested with 5 iterations each | ✅ |
| Performance Range | 0.122 - 3.671 ms | Min to max serialization times | ✅ |

### 📋 Test Scenarios

| Scenario | Status | Serialization | Deserialization |
|----------|--------|---------------|-----------------|
| Web Api Response | ✅ Passed | 1.433 ms | N/A |
| Ml Training Batch | ✅ Passed | 3.671 ms | N/A |
| Financial Transaction | ✅ Passed | 0.18 ms | N/A |
| Mixed Types Challenge | ✅ Passed | 0.251 ms | N/A |
| Security Pii Test | ✅ Passed | 0.122 ms | N/A |

## ℹ️ Baseline Status

No baseline available for comparison. Future PRs will compare against this baseline.

## ✅ Status: Ready for Review

All benchmarks passed! No significant performance regressions detected.

---
*Generated by [datason-benchmarks](https://github.com/danielendler/datason-benchmarks) • Comprehensive Performance Analysis*